requirements.txt
asgiref==3.8.1
attrs==24.2.0
az-iranian-bank-gateways==2.0.10
bleach==6.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-debug-toolbar==4.4.6
django-rosetta==0.10.1
django-summernote==0.8.20.0
idna==3.10
isodate==0.6.1
lxml==5.3.0
pillow==10.4.0
platformdirs==4.3.6
polib==1.2.0
pycryptodome==3.21.0
python-decouple==3.8
pytz==2024.2
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.3
webencodings==0.5.1
zeep==4.2.1
